\name{internals}
\alias{my.rowIQR}
\alias{new.saintoutput}
\alias{saintout}
\alias{TSPM}
\alias{WY.permalg}

\title{
internal functions
}
\description{
Functions for internal usage only
}
\usage{
my.rowIQR(mat)
new.saintoutput(filename)
saintout(test)
TSPM(counts, x1, x0, lib.size, alpha.wh)
WY.permalg(org.statistic, perm.p)
}

\keyword{internal}
